Extract a SearchService from the controller
